address some of the comments

+++ projects/rocsparse/library/src/level3/coomm_device_atomic.h
@@ -84,22 +84,20 @@ namespace rocsparse
                     {
                         for(uint32_t p = 0; p < LOOPS; p++)
                         {
-                            rocsparse::atomic_add(
-                                dense_C,
-                                static_cast<int64_t>((colB + p * WF_SIZE) * ldc + current_row),
-                                dense_C_size,
-                                static_cast<C>(alpha * sum[p]));
+                            rocsparse::atomic_add(dense_C,
+                                                  colB + p * WF_SIZE * ldc + current_row,
+                                                  dense_C_size,
+                                                  static_cast<C>(alpha * sum[p]));
                         }
                     }
                     else
                     {
                         for(uint32_t p = 0; p < LOOPS; p++)
                         {
-                            rocsparse::atomic_add(
-                                dense_C,
-                                static_cast<int64_t>(current_row * ldc + colB + p * WF_SIZE),
-                                dense_C_size,
-                                static_cast<C>(alpha * sum[p]));
+                            rocsparse::atomic_add(dense_C,
+                                                  current_row * ldc + colB + p * WF_SIZE,
+                                                  dense_C_size,
+                                                  static_cast<C>(alpha * sum[p]));
                         }
                     }
 
@@ -137,22 +135,20 @@ namespace rocsparse
             {
                 for(uint32_t p = 0; p < LOOPS; p++)
                 {
-                    rocsparse::atomic_add(
-                        dense_C,
-                        static_cast<int64_t>((colB + p * WF_SIZE) * ldc + current_row),
-                        dense_C_size,
-                        static_cast<C>(alpha * sum[p]));
+                    rocsparse::atomic_add(dense_C,
+                                          colB + p * WF_SIZE * ldc + current_row,
+                                          dense_C_size,
+                                          static_cast<C>(alpha * sum[p]));
                 }
             }
             else
             {
                 for(uint32_t p = 0; p < LOOPS; p++)
                 {
-                    rocsparse::atomic_add(
-                        dense_C,
-                        static_cast<int64_t>(current_row * ldc + colB + p * WF_SIZE),
-                        dense_C_size,
-                        static_cast<C>(alpha * sum[p]));
+                    rocsparse::atomic_add(dense_C,
+                                          current_row * ldc + colB + p * WF_SIZE,
+                                          dense_C_size,
+                                          static_cast<C>(alpha * sum[p]));
                 }
             }
         }
@@ -219,14 +215,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(colB * ldc + current_row),
+                                                  colB * ldc + current_row,
                                                   dense_C_size,
                                                   static_cast<C>(sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(current_row * ldc + colB),
+                                                  current_row * ldc + colB,
                                                   dense_C_size,
                                                   static_cast<C>(sum));
                         }
@@ -287,14 +283,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>((l + swid) * ldc + current_row),
+                                                  l + swid * ldc + current_row,
                                                   dense_C_size,
                                                   static_cast<C>(sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(current_row * ldc + (l + swid)),
+                                                  current_row * ldc + (l + swid),
                                                   dense_C_size,
                                                   static_cast<C>(sum));
                         }
@@ -311,14 +307,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>((l + swid) * ldc + current_row),
+                                                  l + swid * ldc + current_row,
                                                   dense_C_size,
                                                   static_cast<C>(sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(current_row * ldc + (l + swid)),
+                                                  current_row * ldc + (l + swid),
                                                   dense_C_size,
                                                   static_cast<C>(sum));
                         }
@@ -371,14 +367,14 @@ namespace rocsparse
         if(order_C == rocsparse_order_column)
         {
             rocsparse::atomic_add(dense_C,
-                                  static_cast<int64_t>(hipBlockIdx_y * ldc + col),
+                                  hipBlockIdx_y * ldc + col,
                                   dense_C_size,
                                   static_cast<C>(alpha * (val * bval)));
         }
         else
         {
             rocsparse::atomic_add(dense_C,
-                                  static_cast<int64_t>(col * ldc + hipBlockIdx_y),
+                                  col * ldc + hipBlockIdx_y,
                                   dense_C_size,
                                   static_cast<C>(alpha * (val * bval)));
         }
