+++ projects/rocsparse/library/src/level3/coomm_device_segmented_atomic.h
@@ -159,11 +159,11 @@ namespace rocsparse
                     {
                         for(uint32_t p = 0; p < COLS; p++)
                         {
-                            rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(prevrow + (col_offset + p) * ldc
-                                                                   + batch_stride_C * batch),
-                                                  dense_C_size,
-                                                  static_cast<C>(shared_val[p][WF_SIZE - 1]));
+                            rocsparse::atomic_add(
+                                dense_C,
+                                prevrow + (col_offset + p * ldc + batch_stride_C * batch),
+                                dense_C_size,
+                                static_cast<C>(shared_val[p][WF_SIZE - 1]));
                         }
                     }
                     else
@@ -171,8 +171,8 @@ namespace rocsparse
                         for(uint32_t p = 0; p < COLS; p++)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>((col_offset + p) + prevrow * ldc
-                                                                   + batch_stride_C * batch),
+                                                  col_offset + p + prevrow * ldc
+                                                      + batch_stride_C * batch,
                                                   dense_C_size,
                                                   static_cast<C>(shared_val[p][WF_SIZE - 1]));
                         }
@@ -224,11 +224,11 @@ namespace rocsparse
                     {
                         for(uint32_t p = 0; p < COLS; p++)
                         {
-                            rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(row + (col_offset + p) * ldc
-                                                                   + batch_stride_C * batch),
-                                                  dense_C_size,
-                                                  static_cast<C>(val[p]));
+                            rocsparse::atomic_add(
+                                dense_C,
+                                row + (col_offset + p * ldc + batch_stride_C * batch),
+                                dense_C_size,
+                                static_cast<C>(val[p]));
                         }
                     }
                     else
@@ -236,8 +236,8 @@ namespace rocsparse
                         for(uint32_t p = 0; p < COLS; p++)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>((col_offset + p) + row * ldc
-                                                                   + batch_stride_C * batch),
+                                                  col_offset + p + row * ldc
+                                                      + batch_stride_C * batch,
                                                   dense_C_size,
                                                   static_cast<C>(val[p]));
                         }
@@ -253,22 +253,20 @@ namespace rocsparse
             {
                 for(uint32_t p = 0; p < COLS; p++)
                 {
-                    rocsparse::atomic_add(
-                        dense_C,
-                        static_cast<int>(row + (col_offset + p) * ldc + batch_stride_C * batch),
-                        dense_C_size,
-                        static_cast<C>(val[p]));
+                    rocsparse::atomic_add(dense_C,
+                                          row + (col_offset + p * ldc + batch_stride_C * batch),
+                                          dense_C_size,
+                                          static_cast<C>(val[p]));
                 }
             }
             else
             {
                 for(uint32_t p = 0; p < COLS; p++)
                 {
-                    rocsparse::atomic_add(
-                        dense_C,
-                        static_cast<int>((col_offset + p) + row * ldc + batch_stride_C * batch),
-                        dense_C_size,
-                        static_cast<C>(val[p]));
+                    rocsparse::atomic_add(dense_C,
+                                          col_offset + p + row * ldc + batch_stride_C * batch,
+                                          dense_C_size,
+                                          static_cast<C>(val[p]));
                 }
             }
         }
