+++ projects/rocsparse/library/src/level3/csrmm_device_merge.h
@@ -214,22 +214,20 @@ namespace rocsparse
                     {
                         for(uint32_t p = 0; p < LOOPS; p++)
                         {
-                            rocsparse::atomic_add(
-                                dense_C,
-                                static_cast<int64_t>(start_row + ldc * (colB + p * WF_SIZE)),
-                                dense_C_size,
-                                static_cast<C>(alpha * sum[p]));
+                            rocsparse::atomic_add(dense_C,
+                                                  start_row + ldc * (colB + p * WF_SIZE),
+                                                  dense_C_size,
+                                                  static_cast<C>(alpha * sum[p]));
                         }
                     }
                     else
                     {
                         for(uint32_t p = 0; p < LOOPS; p++)
                         {
-                            rocsparse::atomic_add(
-                                dense_C,
-                                static_cast<int>(ldc * start_row + (colB + p * WF_SIZE)),
-                                dense_C_size,
-                                static_cast<C>(alpha * sum[p]));
+                            rocsparse::atomic_add(dense_C,
+                                                  ldc * start_row + (colB + p * WF_SIZE),
+                                                  dense_C_size,
+                                                  static_cast<C>(alpha * sum[p]));
                         }
                     }
                 }
@@ -262,22 +260,20 @@ namespace rocsparse
                     {
                         for(uint32_t p = 0; p < LOOPS; p++)
                         {
-                            rocsparse::atomic_add(
-                                dense_C,
-                                static_cast<int64_t>(start_row + ldc * (colB + p * WF_SIZE)),
-                                dense_C_size,
-                                static_cast<C>(alpha * sum[p]));
+                            rocsparse::atomic_add(dense_C,
+                                                  start_row + ldc * (colB + p * WF_SIZE),
+                                                  dense_C_size,
+                                                  static_cast<C>(alpha * sum[p]));
                         }
                     }
                     else
                     {
                         for(uint32_t p = 0; p < LOOPS; p++)
                         {
-                            rocsparse::atomic_add(
-                                dense_C,
-                                static_cast<int>(ldc * start_row + (colB + p * WF_SIZE)),
-                                dense_C_size,
-                                static_cast<C>(alpha * sum[p]));
+                            rocsparse::atomic_add(dense_C,
+                                                  ldc * start_row + (colB + p * WF_SIZE),
+                                                  dense_C_size,
+                                                  static_cast<C>(alpha * sum[p]));
                         }
                     }
                 }
@@ -358,22 +354,20 @@ namespace rocsparse
                 {
                     for(uint32_t p = 0; p < LOOPS; p++)
                     {
-                        rocsparse::atomic_add(
-                            dense_C,
-                            static_cast<int>(end_row + ldc * (colB + p * WF_SIZE)),
-                            dense_C_size,
-                            static_cast<C>(alpha * sum[p]));
+                        rocsparse::atomic_add(dense_C,
+                                              end_row + ldc * (colB + p * WF_SIZE),
+                                              dense_C_size,
+                                              static_cast<C>(alpha * sum[p]));
                     }
                 }
                 else
                 {
                     for(uint32_t p = 0; p < LOOPS; p++)
                     {
-                        rocsparse::atomic_add(
-                            dense_C,
-                            static_cast<int>(ldc * end_row + (colB + p * WF_SIZE)),
-                            dense_C_size,
-                            static_cast<C>(alpha * sum[p]));
+                        rocsparse::atomic_add(dense_C,
+                                              ldc * end_row + (colB + p * WF_SIZE),
+                                              dense_C_size,
+                                              static_cast<C>(alpha * sum[p]));
                     }
                 }
             }
@@ -460,14 +454,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(start_row + ldc * colB),
+                                                  start_row + ldc * colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(ldc * start_row + colB),
+                                                  ldc * start_row + colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
@@ -495,14 +489,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(start_row + ldc * colB),
+                                                  start_row + ldc * colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(ldc * start_row + colB),
+                                                  ldc * start_row + colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
@@ -566,14 +560,14 @@ namespace rocsparse
                     if(order_C == rocsparse_order_column)
                     {
                         rocsparse::atomic_add(dense_C,
-                                              static_cast<int>(end_row + ldc * colB),
+                                              end_row + ldc * colB,
                                               dense_C_size,
                                               static_cast<C>(alpha * sum));
                     }
                     else
                     {
                         rocsparse::atomic_add(dense_C,
-                                              static_cast<int>(ldc * end_row + colB),
+                                              ldc * end_row + colB,
                                               dense_C_size,
                                               static_cast<C>(alpha * sum));
                     }
@@ -663,14 +657,14 @@ namespace rocsparse
                             if(order_C == rocsparse_order_column)
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(start_row + ldc * colB),
+                                                      start_row + ldc * colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
                             else
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(ldc * start_row + colB),
+                                                      ldc * start_row + colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
@@ -701,14 +695,14 @@ namespace rocsparse
                             if(order_C == rocsparse_order_column)
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(start_row + ldc * colB),
+                                                      start_row + ldc * colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
                             else
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(ldc * start_row + colB),
+                                                      ldc * start_row + colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
@@ -778,14 +772,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(end_row + ldc * colB),
+                                                  end_row + ldc * colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(ldc * end_row + colB),
+                                                  ldc * end_row + colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
@@ -874,14 +868,14 @@ namespace rocsparse
                             if(order_C == rocsparse_order_column)
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(start_row + ldc * colB),
+                                                      start_row + ldc * colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
                             else
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(ldc * start_row + colB),
+                                                      ldc * start_row + colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
@@ -912,14 +906,14 @@ namespace rocsparse
                             if(order_C == rocsparse_order_column)
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(start_row + ldc * colB),
+                                                      start_row + ldc * colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
                             else
                             {
                                 rocsparse::atomic_add(dense_C,
-                                                      static_cast<int>(ldc * start_row + colB),
+                                                      ldc * start_row + colB,
                                                       dense_C_size,
                                                       static_cast<C>(alpha * sum));
                             }
@@ -989,14 +983,14 @@ namespace rocsparse
                         if(order_C == rocsparse_order_column)
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(end_row + ldc * colB),
+                                                  end_row + ldc * colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
                         else
                         {
                             rocsparse::atomic_add(dense_C,
-                                                  static_cast<int>(ldc * end_row + colB),
+                                                  ldc * end_row + colB,
                                                   dense_C_size,
                                                   static_cast<C>(alpha * sum));
                         }
